+++ routes/v3/stats/worldometer.js
@@ -3,15 +3,13 @@ const router = express.Router();
 const asyncHandler = require("express-async-handler");
 const db = require('../../../system/database');
 const cache = require('../../../system/redis-cache');
-const { getCustomStats, fetchDataFromGoogleSheet } = require('../../../services/customStats');
-const { getStatsWithCountryDetail } = require('../../../services/statsService');
 const { cacheCheck } = require('../../../services/cacheMiddleware');
 
  /**
  * @api {get} /v3/stats/worldometer
  * @apiName worldometer
  * @apiGroup Worldometer stats
- * @apiVersion 2.0.0
+ * @apiVersion 3.0.0
  * @apiDescription Returns worldometer stats
  * @apiSuccessExample Response (example):
  * HTTP/1.1 200 Success
@@ -34,6 +32,7 @@ const { cacheCheck } = require('../../../services/cacheMiddleware');
 ]
  */
 router.get('/', cacheCheck, asyncHandler(async function(req, res, next) {
+  console.log('calling v3/worldometer');
   try {
     const result = await getWorldometer();
     return res.json(result);
@@ -44,17 +43,49 @@ router.get('/', cacheCheck, asyncHandler(async function(req, res, next) {
   }
 }));
 
+/**
+ * @api {get} /v3/stats/worldometer/stats_overview
+ * @apiName stats_overview
+ * @apiGroup Get Worldometer overview stats
+ * @apiVersion 3.0.0
+ * @apiDescription Returns stats overview on main and analytics page
+ * @apiSuccessExample Response (example):
+ * HTTP/1.1 200 Success
+  {
+  "totalConfirmed": 276113,
+  "totalDeaths": 11402,
+  "totalRecovered": 91952,
+  "totalNewCases": 562,
+  "totalNewDeaths": 23,
+  "totalActiveCases": 172759,
+  "totalCasesPerMillionPop": 35,
+  "created": "2020-03-21T13:00:13.000Z"
+  }
+ */
+router.get('/stats_overview', cacheCheck, cache.route(), asyncHandler(async function (req, res, next) {
+  console.log('calling v3/worldometer/stats_overview');
+  try {
+    const results = await getStatsOverview();
+    return res.json(results);
+  }
+  catch (error) {
+    console.log('[/stats] error', error);
+    return res.json(error);
+  }
+}));
+
+
 async function getWorldometer() {
   const conn = db.conn.promise();
   let query = `SELECT ac.country_code as countryCode,
-  tt.country, 
-  tt.total_cases AS totalConfirmed, 
-  tt.total_deaths as totalDeaths, 
+  tt.country,
+  tt.total_cases AS totalConfirmed,
+  tt.total_deaths as totalDeaths,
   tt.total_recovered as totalRecovered,
-  tt.new_cases AS dailyConfirmed, 
-  tt.new_deaths AS dailyDeaths, 
-  tt.active_cases as activeCases, 
-  tt.serious_critical_cases AS totalCritical, 
+  tt.new_cases AS dailyConfirmed,
+  tt.new_deaths AS dailyDeaths,
+  tt.active_cases as activeCases,
+  tt.serious_critical_cases AS totalCritical,
   CAST(tt.total_cases_per_million_pop AS UNSIGNED) AS totalConfirmedPerMillionPopulation,
   (tt.total_deaths / tt.total_cases * 100) AS FR,
   (tt.total_recovered / tt.total_cases * 100) AS PR,
@@ -62,7 +93,7 @@ async function getWorldometer() {
         FROM worldometers tt INNER JOIN (
             SELECT country,
             max(last_updated) AS MaxDateTime
-            FROM worldometers tt 
+            FROM worldometers tt
             where country not in ("Sint Maarten","Congo")
            GROUP BY country) groupedtt
       ON tt.country = groupedtt.country
@@ -71,8 +102,27 @@ async function getWorldometer() {
       group by tt.country
       order by tt.total_cases DESC`;
   let result = await conn.query(query);
-  //const result = await conn.query(query);
   return result[0];
+}
+
+async function getStatsOverview() {
+  const conn = db.conn.promise();
+  let query = `
+  SELECT
+total_cases AS totalConfirmed,
+total_deaths as totalDeaths,
+total_recovered as totalRecovered,
+new_cases AS totalNewCases,
+new_deaths AS totalNewDeaths,
+active_cases AS totalActiveCases,
+CAST(total_cases_per_million_pop AS UNSIGNED) AS totalCasesPerMillionPop,
+MAX(last_updated) as created
+FROM worldometers_total_sum
+LIMIT 1
+`;
+
+  let result = await conn.query(query);
+  return result[0][0];
 }
 
 module.exports = router;
